Fix vehicle desync bug when writing to DB
Fix the vehicle desync bug when writing to DB. Each
time player is synced to DB, vehicles are synced at the same time.
Vehicle sync was being done first, which was causing lag to affect
core game synchronization of vehicle inventory contents. Client vehicle
inventory was not synchronized with server's vehicle inventory for a
specific vehicle at the time of saving it to database, which caused
vehicle inventory modifications from clients ineffective if not sitting
directly inside vehicle.

Bug was difficult to track down.