Fix bug: spawn bed deleted after building it
Due to a faulty check on provided mats quantity requiring to be exactly
equal to removed mats quantity after building, the spawn bed is deleted
just after it was built, if player has more mats than needed.

Fix problem by allowing the removed number of mats
be smaller or equal than provided number of items in order to pass the
check.